requirements.txt
seaborn==0.12.2
matplotlib==3.7.0
numpy==1.24.1
scikit-learn==1.2.1
scipy==1.10.1
torch==2.0
torchmetrics==0.11.1
torchvision==0.15.1
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@8523324c116dcf7be6bddb73bf4eb1779ef6e611
nfnets-pytorch==0.1.3
